# Speed up editoast errors db insertion