--- a/editoast/Cargo.toml
+++ b/editoast/Cargo.toml
@@ -60,6 +60,7 @@ serde.workspace = true
 serde_derive.workspace = true
 serde_json.workspace = true
 serde_yaml = "0.9.25"
+sha1 = "0.10"
 strum = "0.25.0"
 strum_macros = "0.25.3"
 thiserror = "1.0.50"