--- a/editoast/migrations/2023-11-24-162712_add_errors_hash/down.sql
+++ b/editoast/migrations/2023-11-24-162712_add_errors_hash/down.sql
@@ -1,0 +1,1 @@
+ALTER TABLE infra_layer_error DROP COLUMN info_hash;