--- a/editoast/src/generated_data/error/sql/routes_insert_errors.sql
+++ b/editoast/src/generated_data/error/sql/routes_insert_errors.sql
@@ -1,14 +1,17 @@
 WITH errors AS (
-    SELECT unnest($2) AS information
+    SELECT unnest($2) AS information,
+        unnest($3) AS error_hash
 )
 INSERT INTO infra_layer_error (
         infra_id,
         geographic,
         schematic,
-        information
+        information,
+        info_hash
     )
 SELECT $1 AS infra_id,
     NULL,
     NULL,
-    errors.information
+    errors.information,
+    error_hash
 FROM errors