## What this PR does

Editing an object takes too much time because we have to refresh the whole error layer each time.
The idea is to identify the new errors and those that have disappeared, once all the new errors have been generated. This avoids deleting everything and inserting all the new errors because in practice there are few changes.
To achieve this, we hash (with sha1) the errors.  

## Benchmark

|                | Ref    | New  |
|----------------|--------|------|
| Move detector  | 4089ms | 11ms |
| Remove a track | 3881ms | 89ms |
| Auto fix       | 4225ms | 30ms |

This benchmark was done using a big infra (France) with ~10,000 errors.
In practice editing a big infra took at least `5s` now it takes `1.5s`.

close #5516