@flomonster 2023-11-24
## Unicity error

I haven't checked why yet but the unicity constraint on the `info_hash` can fail. Two possibilities:
- Our error generator can generate duplicates
- My PR is broken and I'm trying to add an existing error