@flomonster 2023-11-28
I fixed the unicity error. It was not an error in error generators but in the error layer generator.
Operational point errors were duplicated since their geometry is split among multiple rows in DB (needed to have multiple kp). I fixed it using `ST_Collect` so now one error generates only one row in DB.